Allow for generating a blocktemplate and/or submitting a block even if the node is not synced
Note: Generally, this only becomes necessary if the network forks due to non-upgraded mining nodes throwing invalid blocks for consensus purposes. In doing so, the network may believe that there are more blocks than what the local node will accept. By permitting these operations to occur when the node is not synced, it becomes possible for the node to still generate valid blocks for consensus and thus stack the chain higher than the non-upgraded nodes without the need to block the offending peer.